+++ NEWS.md
@@ -1,6 +1,6 @@
 # stdmod 0.1.2 to 0.1.4
 
-- 0.1.4 (WIP)
+- 0.1.5
 
     - Change the column order of the coefficient table
       in `std_selected_boot` output. Bootstrapping confidence
@@ -8,7 +8,8 @@
     - Add `vcov` method for `std_selected` output. If bootstrapping is used,
       it can return the variance-covariance matrix of the bootstrap estimates.
     - Add `confint` method for `std_selected` output. If bootstrapping is used,
-      it can return the bootstrap percentile confidence intervals if requested. 
+      it can return the bootstrap percentile confidence intervals if requested.
+    - Other minor fixes.
 
 - 0.1.3